texture: compressed texture format selection (#1070)
